requirements.txt
Adafruit-Blinka==8.33.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-connectionmanager==1.0.0
adafruit-circuitpython-dht==4.0.4
adafruit-circuitpython-requests==3.0.0
adafruit-circuitpython-typing==1.10.2
Adafruit-PlatformDetect==3.61.0
Adafruit-PureIO==1.1.11
blinker==1.7.0
click==8.1.7
dnspython==2.6.1
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pyftdi==0.55.0
pymongo==4.6.2
pyserial==3.5
python-dotenv==1.0.1
pyusb==1.2.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
sysv-ipc==1.1.0
typing_extensions==4.10.0
Werkzeug==3.0.1
